requirements.txt
SpeechRecognition
edge-tts
pygame
google-generativeai
python-dotenv
pyaudio
